Fix the link timezone activation script

+++ hosts/common/baseline.nix
@@ -43,7 +43,9 @@
     ln -sf ${pkgs.systemd}/share/zsh/site-functions/* /usr/share/zsh/site-functions/
   '';
   system.activationScripts.linkTzdata = ''
-    mkdir -p /usr/share
-    ln -sf ${pkgs.tzdata}/share/zoneinfo /usr/share/zoneinfo
+    if ! [ -L /usr/share/zoneinfo ]; then
+      mkdir -p /usr/share
+      ln -sf ${pkgs.tzdata}/share/zoneinfo /usr/share/zoneinfo
+    fi
   '';
 }
